Move advanced options into their own page, created development build
