+++ apps/harvester/package.json
@@ -22,6 +22,7 @@
     "prod:beta": "yarn run -T cross-env CONFIG_NAME=prodbeta yarn _start:dbg",
     "prod": "yarn run -T cross-env CONFIG_NAME=prod yarn _start:dbg",
     "package": "electron-forge package",
+    "make:dev": "yarn run -T cross-env CONFIG_NAME=development yarn _make",
     "make:prod": "yarn run -T cross-env CONFIG_NAME=prodbeta yarn _make",
     "make:prodtest": "yarn run -T cross-env CONFIG_NAME=prodtest yarn _make",
     "make:validation": "yarn run -T cross-env CONFIG_NAME=prodbeta NOTIFY_ON_MODAL_ENCOUNTER=true yarn _make",
