+++ apps/harvester/src/Advanced/index.module.less
@@ -1,0 +1,3 @@
+.container {
+  padding: 1em;
+}
