+++ apps/harvester/src/results/index.tsx
@@ -1,5 +1,5 @@
 import { useEffect, useState } from 'react'
-import { Button, Checkbox, Dimmer, Loader, Segment } from 'semantic-ui-react'
+import { Button, Dimmer, Loader } from 'semantic-ui-react'
 import { HarvestData } from '../Harvester/types';
 import { fromDb } from '@thecointech/store-harvester';
 import { StateDisplay } from './StateDisplay';
@@ -12,7 +12,6 @@ import styles from './index.module.less';
 export const Results = () => {
 
   const [state, setState] = useState<HarvestData|undefined>();
-  const [visible, setVisible] = useState<boolean>();
   const replayTask = useBackgroundTask("replay");
   const isReplaying = isRunning(replayTask);
 
@@ -30,7 +29,7 @@ export const Results = () => {
   }, [])
   const runImmediately = async () => {
     log.info("Commencing manual run");
-    const r = await window.scraper.runHarvester(!visible);
+    const r = await window.scraper.runHarvester();
     if (r.error) {
       alert("Error - please check logs:\n " + r.error);
     }
@@ -39,84 +38,6 @@ export const Results = () => {
     setState(state.value);
   }
 
-  const exportResults = async () => {
-    const r = await window.scraper.exportResults();
-    if (r.error) {
-      alert("Error - please check logs:\n " + r.error);
-    }
-    const a = window.document.createElement('a');
-    a.href = window.URL.createObjectURL(new Blob([r.value ?? 'no values'], { type: 'text/csv' }));
-    a.download = 'results.csv';
-
-    // Append anchor to body.
-    document.body.appendChild(a);
-    a.click();
-
-    // Remove anchor from body
-    document.body.removeChild(a);
-    window.URL.revokeObjectURL(a.href);
-  }
-
-  const exportConfig = async () => {
-    const r = await window.scraper.exportConfig();
-    if (r.error) {
-      alert("Error - please check logs:\n " + r.error);
-    }
-    const a = window.document.createElement('a');
-    a.href = window.URL.createObjectURL(new Blob([r.value ?? 'no values'], { type: 'text/csv' }));
-    a.download = 'config.json';
-
-    // Append anchor to body.
-    document.body.appendChild(a);
-    a.click();
-
-    // Remove anchor from body
-    document.body.removeChild(a);
-    window.URL.revokeObjectURL(a.href);
-  }
-
-  const importConfig = async () => {
-    // Create file input element
-    const input = document.createElement('input');
-    input.type = 'file';
-    input.accept = '.json';
-
-    // Handle file selection
-    input.onchange = async (e) => {
-      const file = (e.target as HTMLInputElement).files?.[0];
-      if (!file) return;
-
-      try {
-        // Read and parse the JSON file
-        const text = await file.text();
-        const config = JSON.parse(text);
-
-        // Validate that we can get a scraping config
-        // getScrapingScript(config);
-        // Set the process config
-        const r = await window.scraper.importScraperScript(config);
-        if (r.error) {
-          alert("Error - please check logs:\n " + r.error);
-        }
-        else {
-          alert("Config imported successfully!");
-        }
-      } catch (err) {
-        console.error('Failed to import config:', err);
-        alert('Failed to import configuration file. Please ensure it is a valid JSON file with scraping settings.');
-      }
-    };
-
-    // Trigger file selection
-    input.click();
-  }
-
-  async function launchBrowser() {
-    const r = await window.scraper.warmup("_blank");
-    if (r.error) {
-      alert("Error - please check logs:\n " + r.error);
-    }
-  }
 
   return (
     <div id={styles.container}>
@@ -125,21 +46,8 @@ export const Results = () => {
           <Loader>Running</Loader>
         </Dimmer>
         <StateDisplay state={state} />
-        {/* <div>
-          <Button onClick={launchBrowser}>Launch Browser</Button>
-        </div> */}
-        <div>
-          <Button onClick={exportResults}>Export Results</Button>
-        </div>
-        {/* <div>
-          <Button onClick={exportConfig}>Export Config</Button>
-        </div> */}
-        {/* <div>
-          <Button onClick={importConfig}>Import Script</Button>
-        </div> */}
         <div>
           <Button onClick={runImmediately}>Run Harvester Now</Button>
-          <Checkbox onClick={(_, {checked}) => setVisible(checked)} checked={visible} label="Override Visibility" />
         </div>
       </Dimmer.Dimmable>
       <BackgroundTaskProgressBar type='replay' />
